Adapt to previous changes
